Modify object creation, name binding, and typelog
We don't need to set the name when we create an object.  It causes
an annoyng scenario where the name can't contain the object id.
Now names are set independently of object creation.

Bound names can be updated more easily (and accurately) now.

All the silly type keeping is removed.  It ended up being a list of
auto-increment ids, but that's just... 1...n.